refactor: extract violation_age into analyzer/violation_age.rs (#190)
Moves ViolationAgeSummary + compute_violation_age (plus their two tests)
out of the 3000-line analyzer/mod.rs into a dedicated submodule. mod.rs
re-exports both via `pub use` so callers see no change.
